fix(checker): TS2589 no longer bundled with TS2615 for circular mapped-type aliases
When a type-reference application detects a self-referencing mapped-type
alias body (circular_mapped), tsc emits TS2615 (property circularly
references itself) as a signal independent of TS2589 (excessively deep
instantiation) -- the two only co-occur when the instantiation depth guard
actually trips. tsz unconditionally emitted TS2589 alongside TS2615
whenever circular_mapped was true, regardless of the depth guard, which is
the conformance false positive on
TypeScript/tests/cases/compiler/recursivelyExpandingUnionNoStackoverflow.ts
(`type N<T, K extends string> = T | { [P in K]: N<T, K> }[K];`).

Verified against pinned typescript@7.0.2: only TS2615 fires on that
fixture and two renamed-binder/multi-instantiation variants; a genuinely
excessively-deep non-mapped alias (existing ts2589_message_text repro)
still reports TS2589 alone, and the keyof-constraint case still reports
TS2589 alone with no TS2615 (existing negative control).

Splits the two diagnostics' emission at both call sites in
type_resolution/core.rs so TS2589/tuple-too-large is gated on the depth
guard (`exceeded`/`tuple_too_large`) alone, while TS2615 stays gated on
`circular_mapped` independently.

Rewrites the existing unit test that asserted the wrong, never
oracle-verified bundled behavior, and adds a renamed-binder/multi-site
case plus a negative control confirming genuine deep recursion without a
circular mapped body still reports TS2589 alone.

Verification:
- Release binary against the real conformance fixture and pinned
  typescript@7.0.2 oracle: exact match (TS2615 only).
- cargo test -p tsz-checker --lib ts2589: all tests green (see PR body for
  count).

Split the mapped-type-related tests (circular_mapped_type_alias_*,
homomorphic_self_mapped_*, and 4 adjacent recursive-mapped-type cases) out
of ts2589_tests.rs into a new tests/ts2589_mapped_type_tests.rs, registered
via #[path] in src/lib.rs alongside the existing module -- the new tests
pushed ts2589_tests.rs over the arch-size 2000-line test-file cap (#16745).
70/70 ts2589 tests pass across both files after the split.
